feat: Add input to select the stations CSV
