+++ app/assets/translations.json
@@ -45,6 +45,7 @@
         "title": "Detect animals",
         "chooseInput": "Choose directory with photos",
         "chooseOutput": "Choose directory to save classification results",
+        "chooseStationsCsv": "Chose file with station coordinates (optional)",
         "chooseModel": "Select AI model to use:",
         "projectId": "Project ID (optional)",
         "deploymentId": "Deployment ID (optional)",
@@ -154,6 +155,7 @@
         "title": "Détectez animaux",
         "chooseInput": "Choisissez un dossier avec des photos",
         "chooseOutput": "Choisissez un dossier pour sauvegarder les résultats de la classification",
+        "chooseStationsCsv": "TODO",
         "chooseModel": "Sélectionnez le modèle d'IA à utiliser :",
         "projectId": "ID du projet",
         "deploymentId": "ID de déploiement",
